Organize types and sorting functions
