+++ packages/ui-components/src/components/CompareTable/index.ts
@@ -2,4 +2,4 @@ export * from "./CompareTable";
 export * from "./CompareTable.style";
 export * from "./SortControls";
 export * from "./ColumnHeader";
-export * from "./interfaces";
+export * from "./utils";
