+++ packages/ui-components/src/components/MetricCompareTable/utils.tsx
@@ -2,17 +2,12 @@ import React from "react";
 import isNumber from "lodash/isNumber";
 import { Stack, Typography } from "@mui/material";
 import { Metric, MultiMetricDataStore } from "@actnowcoalition/metrics";
-import { RegionDB } from "@actnowcoalition/regions";
+import { Region, RegionDB } from "@actnowcoalition/regions";
 import { formatPopulation } from "../../common/utils";
+import { SortDirection } from "../../common/utils/compare";
 import { MetricValue } from "../MetricValue";
-import {
-  ColumnDefinition,
-  ColumnHeader,
-  SortDirection,
-  getAriaSort,
-} from "../CompareTable";
+import { ColumnDefinition, ColumnHeader, getAriaSort } from "../CompareTable";
 import { StyledTableCell, StyledLink } from "./MetricCompareTable.style";
-import { Region } from "@actnowcoalition/regions";
 
 export interface Row {
   /** Unique ID for the row. */
